feat(browser): make Firefox control available without desktop
