+++ apps/app-desktop/package.json
@@ -20,15 +20,14 @@
   },
   "devDependencies": {
     "@types/node": "^20.19.39",
-    "@types/ws": "^8.5.13",
     "electron": "^43.2.0",
     "electron-builder": "^25.0.0",
     "typescript": "^5.7",
     "vitest": "^3"
   },
   "dependencies": {
+    "@use-brian/firefox-companion": "workspace:*",
     "electron-updater": "^6.8.9",
-    "ws": "^8.18.0",
     "zod": "^3.24"
   }
 }
