+++ apps/browser-extension/CLAUDE.md
@@ -17,7 +17,7 @@ browser-relay (`apps/browser-relay`); the app-web connect surface is
 - `pairing.ts` — the credential transition behind every Connect, plus the `externally_connectable` sender check (`[COMP:ext/pairing]`).
 - `snapshot.ts` — CDP accessibility tree into the shared snapshot shape.
 - `popup.ts` / `popup-status.ts` / `allow.ts` — the pairing popup, its status wording, and the per-task allow prompt.
-- `firefox-background.ts` / `firefox-native-client.ts` — Firefox relay/consent wiring and the fixed-operation native-messaging bridge to the installed Electron companion (`[COMP:ext/firefox-agent]`).
+- `firefox-background.ts` / `firefox-native-client.ts` — Firefox relay/consent wiring and the fixed-operation native-messaging bridge to either the standalone companion or the installed Electron app (`[COMP:ext/firefox-agent]`).
 - `firefox-popup.ts` / `firefox-allow.ts` — the Firefox pages, visually matched to the Chromium popup/allow pages.
 
 ## Governance guardrail (my-browser.md §4 D4 / §6) — DO NOT WIDEN
@@ -30,9 +30,11 @@ The narrow surface **is** the feature. The extension:
 
 Firefox has no WebExtension debugger API. Its build therefore requests only
 `nativeMessaging`, `tabs`, and `storage`, and delegates the same fixed operation
-vocabulary to the Use Brian Electron executable over the registered
-`ai.usebrian.browser` host. The companion connects to Firefox's loopback-only
-WebDriver BiDi endpoint; Firefox must have been started through the desktop app.
+vocabulary to the registered `ai.usebrian.browser` native host. The companion
+connects to Firefox's loopback-only WebDriver BiDi endpoint; Firefox must have
+been started through either `use-brian-firefox start` or the desktop app. The
+Electron-free implementation lives in `apps/firefox-companion` and is shared by
+both hosts (`[COMP:ext/firefox-companion]`).
 This is not permission to add content scripts or host access. The Firefox manifest
 test locks the same no-host/no-content-script boundary.
