+++ apps/browser-extension/static-firefox/popup.html
@@ -25,7 +25,7 @@
     <div id="status" class="status"><span class="dot"></span><span id="status-text">Checking...</span></div>
     <div id="build-warning" class="warn" hidden></div>
     <div class="row" id="grant-row" hidden>
-      <button id="grant" class="primary" style="width: 100%">Open Use Brian desktop setup</button>
+      <button id="grant" class="primary" style="width: 100%">Open Firefox setup</button>
     </div>
     <div class="row">
       <label for="relay-url">Relay URL</label>
