+++ apps/firefox-companion/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./firefox-bidi.js";
+export * from "./firefox-launcher.js";
+export * from "./firefox-native-host.js";
+export * from "./firefox-native-registration.js";
