+++ packages/shared/src/browser-extension-build.ts
@@ -19,7 +19,7 @@
  * Do not hand-edit. `pnpm check` (`invariants/browser-extension-build-stamp`)
  * prints the value to paste when extension source moves.
  */
-export const CURRENT_EXTENSION_BUILD = '70dd0b6751ad'
+export const CURRENT_EXTENSION_BUILD = '31dec9925224'
 
 /**
  * A reported build is stale unless it matches exactly.
